fix: Refactor code in messageUpdate.js file.
The function `showResult()` should be outside the main method, because it will mean, that every time the main method (`messageUpdate`) is executed, it will create and execute the function `showResult`. From now on, the function is defined in the first import of the file and executed when `messageUpdate` method is called.